+++ CHANGELOG.md
@@ -9,8 +9,11 @@ Multi-planar capture will not be targeted in the near future unless someone else
 
 
 ## [0.13.0]
+### Added
+- Handling of boolean and button controls
 ### Changed
 - Simplified examples, removing clap argument parsing
+- Unified Value/Value64 control types into a single Integer enum variant
 
 ## [0.12.1] - 2020-05-01
 ### Fixed